requirements.txt
pandas
nltk
tqdm
python-Levenshtein